Crate itermore

🤸‍♀️ More iterator adaptors.

This crate provides some useful iterator adaptors and functions. Unlike itertools this crate provides a separate extension trait for each adaptor. Additionally, each type of adaptor is feature flagged so you only have to compile the features you need.

Getting started

Add the crate to Cargo manifest.

cargo add itermore --features full

And bring the extension traits into scope.

use itermore::prelude::*;

Now you can use extension methods like array_windows on any iterator.

for [a, b, c] in iter.array_windows() {
    println!("{} {} {}", a, b, c)
}
// Outputs
//    1 2 3
//    2 3 4
//    3 4 5

It is recommended to only enable the features that you need. For example if you only want the array_combinations adaptor you would add the following to your Cargo manifest.

[dependencies]
itermore = { version = "*", features = ["array_combinations"]}

Provided functionality

Methods

collect_array: Collects an iterator into an array.

min_max and friends: Returns the minimum and maximum element of an iterator.

next_chunk: Returns the next N elements of the iterator as an array.

sorted and friends: Returns a new iterator with all elements sorted.

Adaptors

array_chunks returns an iterator over N elements of the iterator at a time.

array_windows returns an iterator over all contiguous windows of length N.

array_combinations returns an iterator over K length combinations of all the elements in the underlying iterator.

array_combinations_with_reps returns an iterator over K length combinations with repetitions/replacements of all the elements in the underlying iterator.

cartesian_product returns an iterator over the cartesian product of the element sets of two iterators.

circular_array_windows returns an iterator over all contiguous windows of length N that wraps around at the end.

combinations returns an iterator over k length combinations of all the elements in the underlying iterator.

combinations_with_reps returns an iterator over k length combinations with repetitions/replacements of all the elements in the underlying iterator.

Modules

prelude

Re-exports all iterator extension traits.

Macros

cartesian_product

Returns an iterator over the cartesian product of the element sets of multiple iterators (up to 12).

Structs

ArrayChunks

An iterator over N elements of the iterator at a time.

ArrayCombinations

An iterator that iterates over K length combinations of all the elements in the underlying iterator.

ArrayCombinationsWithReps

An iterator that iterates over K length combinations with repetitions/replacements of all the elements in the underlying iterator.

ArrayWindows

An iterator over all contiguous windows of length N.

CartesianProduct

An iterator over the cartesian product of the element sets of two iterators I and J.

CircularArrayWindows

An iterator over all contiguous windows of length N wrapping back to the first elements when the window would otherwise exceed the length of the iterator.

Combinations

An iterator that iterates over k length combinations of all the elements in the underlying iterator.

CombinationsWithReps

An iterator that iterates over K length combinations with repetitions/replacements of all the elements in the underlying iterator.

Traits

IterArrayChunks

An extension trait that provides the array_chunks method for iterators.

IterArrayCombinations

An extension trait that provides the array_combinations method for iterators.

IterArrayCombinationsWithReps

An extension trait that provides the array_combinations_with_reps method for iterators.

IterArrayWindows

An extension trait that provides the array_windows method for iterators.

IterCartesianProduct

An extension trait that provides the cartesian_product method for iterators.

IterCircularArrayWindows

IterCollectArray

An extension trait that provides the collect_array method for iterators.

IterCombinations

An extension trait that provides the combinations method for iterators.

IterCombinationsWithReps

An extension trait that provides the combinations_with_reps method for iterators.

IterMinMax

An extension trait that provides the min_max method and friends for iterators.

IterNextChunk

An extension trait that provides the next_chunk method for iterators.

IterSorted

An extension trait that provides the sorted method and friends for iterators.