itermore

Module prelude

Re-exports all iterator extension traits.

The intention is that this module is used as a * import.

use itermore::prelude::*;

If you want to refer to a trait directly rather import it from the crate root.

Re-exports

pub use super::IterArrayChunks;

pub use super::IterArrayCombinations;

pub use super::IterArrayCombinationsWithReps;

pub use super::IterArrayWindows;

pub use super::IterCartesianProduct;

pub use super::IterCircularArrayWindows;

pub use super::IterCollectArray;

pub use super::IterCombinations;

pub use super::IterCombinationsWithReps;

pub use super::IterMinMax;

pub use super::IterNextChunk;

pub use super::IterSorted;