itermore

Trait IterCartesianProduct

pub trait IterCartesianProduct: Iterator {
    // Provided method
    fn cartesian_product<J>(
        self,
        other: J,
    ) -> CartesianProduct<Self, J::IntoIter> 
       where Self: Sized,
             Self::Item: Clone,
             J: IntoIterator,
             J::IntoIter: Clone { ... }
}

An extension trait that provides the cartesian_product method for iterators.

Provided Methods

fn cartesian_product<J>(self, other: J) -> CartesianProduct<Self, J::IntoIter>
where Self: Sized, Self::Item: Clone, J: IntoIterator, J::IntoIter: Clone,

Returns an iterator adaptor that iterates over the cartesian product of the element sets of two iterators self and other.into_iter().

Examples

use itermore::IterCartesianProduct;

let v = Vec::from_iter((0..3).cartesian_product("αβ".chars()));
assert_eq!(v, [(0, 'α'), (0, 'β'), (1, 'α'), (1, 'β'), (2, 'α'), (2, 'β')]);